move pagination inside data_table file
