+++ src/javascript/app_2/pages/trading/components/data_table.jsx
@@ -1,50 +1,188 @@
 import React from 'react';
-import Pagination from './pagination';
 
-// generate dummy data
-const transactions = Array(150).fill(0).map((_, i) => {
-    return {
-        balance_after: 10150.1300,
-        transaction_id: (10867502908 - i),
-        reference_id: (45143958928 - i),
-        transaction_time: (1441175849 - i * 100),
-        action_type: (i % 2 ? 'Sell' : 'Buy'),
-        amount: -83.2300,
-        longcode: 'Win payout if the last digit of Volatility 25 Index is 7 after 5 ticks.',
-        payout: 90.91
-    };
-});
-
-const statement_columns = [
-    {
-        title: 'Date',
-        dataIndex: 'transaction_time'
-    },
-    {
-        title: 'Ref.',
-        dataIndex: 'reference_id'
-    },
-    {
-        title: 'Potential payout',
-        dataIndex: 'payout'
-    },
-    {
-        title: 'Action',
-        dataIndex: 'action_type'
-    },
-    {
-        title: 'Description',
-        dataIndex: 'longcode'
-    },
-    {
-        title: 'Credit/Debit',
-        dataIndex: 'amount'
-    },
-    {
-        title: 'Balance (USD)',
-        dataIndex: 'balance_after'
+class Pagination extends React.PureComponent {
+    constructor(props) {
+        super(props);
+        this.handleChange = this.handleChange.bind(this);
+        this.calcNumOfPages = this.calcNumOfPages.bind(this);
+        this.handleNext = this.handleNext.bind(this);
+        this.handlePrev = this.handlePrev.bind(this);
+        this.handleJumpUp = this.handleJumpUp.bind(this);
+        this.handleJumpDown = this.handleJumpDown.bind(this);
+        this.renderUpEllipsis = this.renderUpEllipsis.bind(this);
+        this.renderDownEllipsis = this.renderDownEllipsis.bind(this);
+        this.renderItem = this.renderItem.bind(this);
+        this.renderItemRange = this.renderItemRange.bind(this);
+        this.renderItems = this.renderItems.bind(this);
+        this.state = {
+            current: 1
+        };
     }
-];
+
+    handleChange(newPage) {
+        if (newPage === this.state.current) return;
+
+        const { pageSize } = this.props;
+
+        this.setState({
+            current: newPage
+        });
+
+        this.props.onChange(newPage, pageSize);
+    }
+
+    calcNumOfPages() {
+        const { total, pageSize } = this.props;
+        return Math.ceil(total / pageSize);
+    }
+
+    handleNext() {
+        if (this.state.current < this.calcNumOfPages()) {
+            this.handleChange(this.state.current + 1);
+        }
+    }
+
+    handlePrev() {
+        if (this.state.current > 1) {
+            this.handleChange(this.state.current - 1);
+        }
+    }
+
+    handleJumpUp() {
+        this.handleChange(Math.min(
+            this.state.current + 5,
+            this.calcNumOfPages()
+        ));
+    }
+
+    handleJumpDown() {
+        this.handleChange(Math.max(
+            1,
+            this.state.current - 5
+        ));
+    }
+
+    renderUpEllipsis() {
+        return (
+            <li
+                className='pagination-item pagination-ellipsis'
+                key='ellipsis-up'
+                onClick={this.handleJumpUp}
+            >
+                <a>...</a>
+            </li>
+        );
+    }
+
+    renderDownEllipsis() {
+        return (
+            <li
+                className='pagination-item pagination-ellipsis'
+                key='ellipsis-down'
+                onClick={this.handleJumpDown}
+            >
+                <a>...</a>
+            </li>
+        );
+    }
+
+    renderItem(pageNum) {
+        return (
+            <li
+                className={`pagination-item ${pageNum === this.state.current ? 'pagination-item-active' : ''}`}
+                key={pageNum}
+                onClick={() => {
+                    this.handleChange(pageNum)
+                }}
+            >
+                <a>{pageNum}</a>
+            </li>
+        );
+    }
+
+    renderItemRange(first, last) {
+        const items = [];
+
+        for (let pageNum = first; pageNum <= last; pageNum++) {
+            items.push(this.renderItem(pageNum));
+        }
+        return items;
+    }
+
+    renderItems() {
+        const numOfPages = this.calcNumOfPages();
+        const { current } = this.state;
+
+        if (numOfPages <= 9) {
+            return this.renderItemRange(1, numOfPages);
+        }
+        else if (current <= 3) {
+            return [
+                ...this.renderItemRange(1, 5),
+                this.renderUpEllipsis(),
+                this.renderItem(numOfPages)
+            ];
+        }
+        else if (current === 4) {
+            return [
+                ...this.renderItemRange(1, 6),
+                this.renderUpEllipsis(),
+                this.renderItem(numOfPages)
+            ];
+        }
+        else if (current === numOfPages - 3) {
+            return [
+                this.renderItem(1),
+                this.renderDownEllipsis(),
+                ...this.renderItemRange(numOfPages - 5, numOfPages)
+            ];
+        }
+        else if (numOfPages - current < 3) {
+            return [
+                this.renderItem(1),
+                this.renderDownEllipsis(),
+                ...this.renderItemRange(numOfPages - 4, numOfPages)
+            ];
+        }
+        else {
+            return [
+                this.renderItem(1),
+                this.renderDownEllipsis(),
+                ...this.renderItemRange(current - 2, current + 2),
+                this.renderUpEllipsis(),
+                this.renderItem(numOfPages)
+            ];
+        }
+    }
+
+    render() {
+        const { current } = this.state;
+        return (
+            <ul className='pagination'>
+                <li
+                    className={`pagination-prev ${current === 1 ? 'pagination-disabled' : ''}`}
+                    onClick={this.handlePrev}
+                >
+                    <a>&lt;</a>
+                </li>
+                {this.renderItems()}
+                <li
+                    className={`pagination-next ${current === this.calcNumOfPages() ? 'pagination-disabled' : ''}`}
+                    onClick={this.handleNext}
+                >
+                    <a>&gt;</a>
+                </li>
+            </ul>
+        );
+    }
+}
+
+Pagination.defaultProps = {
+    total: 0,
+    pageSize: 10,
+    onChange: (page, pageSize) => {console.log(page, pageSize)}
+};
+
 
 class DataTable extends React.Component {
     constructor(props) {
@@ -114,10 +252,7 @@ class DataTable extends React.Component {
 }
 
 DataTable.defaultProps = {
-    dataSource: transactions,
-    columns: statement_columns,
-    pagination: true,
-    pageSize: 6
+    pagination: true
 };
 
 export default DataTable;
