+++ src/templates/static/new_layout/style_guide.jsx
@@ -2,7 +2,52 @@ import React from 'react';
 import Button from '../../../javascript/app_2/pages/trading/components/form/button.jsx';
 import InputField from '../../../javascript/app_2/pages/trading/components/form/input_field.jsx';
 import DataTable from '../../../javascript/app_2/pages/trading/components/data_table.jsx';
-import Pagination from '../../../javascript/app_2/pages/trading/components/pagination.jsx';
+
+// generate dummy data for the DataTable
+const transactions = Array(150).fill(0).map((_, i) => {
+    return {
+        balance_after: 10150.1300,
+        transaction_id: (10867502908 - i),
+        reference_id: (45143958928 - i),
+        transaction_time: (1441175849 - i * 100),
+        action_type: (i % 2 ? 'Sell' : 'Buy'),
+        amount: -83.2300,
+        longcode: 'Win payout if the last digit of Volatility 25 Index is 7 after 5 ticks.',
+        payout: 90.91
+    };
+});
+
+const statement_columns = [
+    {
+        title: 'Date',
+        dataIndex: 'transaction_time'
+    },
+    {
+        title: 'Ref.',
+        dataIndex: 'reference_id'
+    },
+    {
+        title: 'Potential payout',
+        dataIndex: 'payout'
+    },
+    {
+        title: 'Action',
+        dataIndex: 'action_type'
+    },
+    {
+        title: 'Description',
+        dataIndex: 'longcode'
+    },
+    {
+        title: 'Credit/Debit',
+        dataIndex: 'amount'
+    },
+    {
+        title: 'Balance (USD)',
+        dataIndex: 'balance_after'
+    }
+];
+
 
 const StyleGuide = () => (
     <div className='container'>
@@ -101,15 +146,7 @@ const StyleGuide = () => (
             <div className='gr-12'>
                 <h2 className='center-text'>Data Table</h2>
 
-                <DataTable />
-            </div>
-        </div>
-
-        <div className='gr-row gr-padding-20'>
-            <div className='gr-12'>
-                <h2 className='center-text'>Pagination</h2>
-
-                <Pagination total={50} pageSize={10} />
+                <DataTable dataSource={transactions} columns={statement_columns} pagination pageSize={6} />
             </div>
         </div>
     </div>
